+++ controller/handlers.go
@@ -73,7 +73,7 @@ func (c *controller) cmdHandler() {
 
 	case sd.prefix + "clear":
 		c.msgType = "none"
-		err = sd.clear.ClearHandler(req, msg)
+		err = sd.Clear.ClearHandler(req, msg)
 
 	case sd.prefix + "crypto":
 		c.response, err = sd.crypto.Game(req, msg)